Move code out of driver.
