+++ packages/driver-typescript/package.json
@@ -45,7 +45,6 @@
     "typescript": "^4.0.0"
   },
   "dependencies": {
-    "@boost/event": "^2.3.1",
     "rimraf": "^3.0.2"
   },
   "tsconfig": {
